dist-cc compid in GPU back to uint32_t

+++ dist_apps/cc_pull/gen_cuda.cu
@@ -5,11 +5,11 @@
 void kernel_sizing(CSRGraph &, dim3 &, dim3 &);
 #define TB_SIZE 256
 const char *GGC_OPTIONS = "coop_conv=False $ outline_iterate_gb=False $ backoff_blocking_factor=4 $ parcomb=True $ np_schedulers=set(['fg', 'tb', 'wp']) $ cc_disable=set([]) $ hacks=set([]) $ np_factor=8 $ instrument=set([]) $ unroll=[] $ instrument_mode=None $ read_props=None $ outline_iterate=True $ ignore_nested_errors=False $ np=True $ write_props=None $ quiet_cgen=True $ retry_backoff=True $ cuda.graph_type=basic $ cuda.use_worklist_slots=True $ cuda.worklist_type=basic";
-unsigned long long * P_COMP_CURRENT;
+uint32_t * P_COMP_CURRENT;
 #include "kernels/reduce.cuh"
 #include "gen_cuda.cuh"
 static const int __tb_ConnectedComp = TB_SIZE;
-__global__ void InitializeGraph(CSRGraph graph, unsigned int __nowned, unsigned int __begin, unsigned int __end, unsigned long long * p_comp_current)
+__global__ void InitializeGraph(CSRGraph graph, unsigned int __nowned, unsigned int __begin, unsigned int __end, uint32_t * p_comp_current)
 {
   unsigned tid = TID_1D;
   unsigned nthreads = TOTAL_THREADS_1D;
@@ -27,7 +27,7 @@ __global__ void InitializeGraph(CSRGraph graph, unsigned int __nowned, unsigned 
   }
   // FP: "7 -> 8;
 }
-__global__ void ConnectedComp(CSRGraph graph, DynamicBitset *is_updated, unsigned int __nowned, unsigned int __begin, unsigned int __end, unsigned long long * p_comp_current, HGAccumulator<int> ret_val)
+__global__ void ConnectedComp(CSRGraph graph, DynamicBitset *is_updated, unsigned int __nowned, unsigned int __begin, unsigned int __end, uint32_t * p_comp_current, HGAccumulator<int> ret_val)
 {
   unsigned tid = TID_1D;
   unsigned nthreads = TOTAL_THREADS_1D;
@@ -111,8 +111,8 @@ __global__ void ConnectedComp(CSRGraph graph, DynamicBitset *is_updated, unsigne
         jj = ns +_np_j;
         {
           index_type dst;
-          unsigned long long new_comp;
-          unsigned long long old_comp;
+          uint32_t new_comp;
+          uint32_t old_comp;
           dst = graph.getAbsDestination(jj);
           new_comp = p_comp_current[dst];
           if (p_comp_current[src] > new_comp) {
@@ -155,8 +155,8 @@ __global__ void ConnectedComp(CSRGraph graph, DynamicBitset *is_updated, unsigne
           jj = _np_w_start +_np_ii;
           {
             index_type dst;
-            unsigned long long new_comp;
-            unsigned long long old_comp;
+            uint32_t new_comp;
+            uint32_t old_comp;
             dst = graph.getAbsDestination(jj);
             new_comp = p_comp_current[dst];
             if (p_comp_current[src] > new_comp) {
@@ -190,8 +190,8 @@ __global__ void ConnectedComp(CSRGraph graph, DynamicBitset *is_updated, unsigne
         jj= nps.fg.itvalue[_np_i];
         {
           index_type dst;
-          unsigned long long new_comp;
-          unsigned long long old_comp;
+          uint32_t new_comp;
+          uint32_t old_comp;
           dst = graph.getAbsDestination(jj);
           new_comp = p_comp_current[dst];
           if (p_comp_current[src] > new_comp) {
@@ -212,7 +212,7 @@ __global__ void ConnectedComp(CSRGraph graph, DynamicBitset *is_updated, unsigne
   }
   ret_val.thread_exit<_br>(_ts);
 }
-__global__ void ConnectedCompSanityCheck(CSRGraph graph, unsigned int __begin, unsigned int __end, unsigned long long * p_comp_current, HGAccumulator<unsigned int> sum)
+__global__ void ConnectedCompSanityCheck(CSRGraph graph, unsigned int __begin, unsigned int __end, uint32_t * p_comp_current, HGAccumulator<unsigned int> sum)
 {
   unsigned tid = TID_1D;
   unsigned nthreads = TOTAL_THREADS_1D;
