+++ dist_apps/cc_push/gen_cuda.cu
@@ -5,13 +5,13 @@
 void kernel_sizing(CSRGraph &, dim3 &, dim3 &);
 #define TB_SIZE 256
 const char *GGC_OPTIONS = "coop_conv=False $ outline_iterate_gb=False $ backoff_blocking_factor=4 $ parcomb=True $ np_schedulers=set(['fg', 'tb', 'wp']) $ cc_disable=set([]) $ hacks=set([]) $ np_factor=8 $ instrument=set([]) $ unroll=[] $ instrument_mode=None $ read_props=None $ outline_iterate=True $ ignore_nested_errors=False $ np=True $ write_props=None $ quiet_cgen=True $ retry_backoff=True $ cuda.graph_type=basic $ cuda.use_worklist_slots=True $ cuda.worklist_type=basic";
-unsigned long long * P_COMP_CURRENT;
-unsigned long long * P_COMP_OLD;
+uint32_t * P_COMP_CURRENT;
+uint32_t * P_COMP_OLD;
 #include "kernels/reduce.cuh"
 #include "gen_cuda.cuh"
 static const int __tb_ConnectedComp = TB_SIZE;
 static const int __tb_FirstItr_ConnectedComp = TB_SIZE;
-__global__ void InitializeGraph(CSRGraph graph, unsigned int __nowned, unsigned int __begin, unsigned int __end, unsigned long long * p_comp_current, unsigned long long * p_comp_old)
+__global__ void InitializeGraph(CSRGraph graph, unsigned int __nowned, unsigned int __begin, unsigned int __end, uint32_t * p_comp_current, uint32_t * p_comp_old)
 {
   unsigned tid = TID_1D;
   unsigned nthreads = TOTAL_THREADS_1D;
@@ -30,7 +30,7 @@ __global__ void InitializeGraph(CSRGraph graph, unsigned int __nowned, unsigned 
   }
   // FP: "8 -> 9;
 }
-__global__ void FirstItr_ConnectedComp(CSRGraph graph, DynamicBitset *is_updated, unsigned int __nowned, unsigned int __begin, unsigned int __end, unsigned long long * p_comp_current, unsigned long long * p_comp_old)
+__global__ void FirstItr_ConnectedComp(CSRGraph graph, DynamicBitset *is_updated, unsigned int __nowned, unsigned int __begin, unsigned int __end, uint32_t * p_comp_current, uint32_t * p_comp_old)
 {
   unsigned tid = TID_1D;
   unsigned nthreads = TOTAL_THREADS_1D;
@@ -136,11 +136,11 @@ __global__ void FirstItr_ConnectedComp(CSRGraph graph, DynamicBitset *is_updated
         jj = ns +_np_j;
         {
           index_type dst;
-          unsigned long long new_dist;
+          uint32_t new_dist;
           dst = graph.getAbsDestination(jj);
           new_dist = p_comp_current[src];
           if (p_comp_current[dst] > new_dist) {
-            unsigned long long old_dist = atomicMin(&p_comp_current[dst], new_dist);
+            uint32_t old_dist = atomicMin(&p_comp_current[dst], new_dist);
             if (old_dist > new_dist) is_updated->set(dst);
           }
         }
@@ -180,11 +180,11 @@ __global__ void FirstItr_ConnectedComp(CSRGraph graph, DynamicBitset *is_updated
           jj = _np_w_start +_np_ii;
           {
             index_type dst;
-            unsigned long long new_dist;
+            uint32_t new_dist;
             dst = graph.getAbsDestination(jj);
             new_dist = p_comp_current[src];
             if (p_comp_current[dst] > new_dist) {
-              unsigned long long old_dist = atomicMin(&p_comp_current[dst], new_dist);
+              uint32_t old_dist = atomicMin(&p_comp_current[dst], new_dist);
               if (old_dist > new_dist) is_updated->set(dst);
             }
           }
@@ -220,11 +220,11 @@ __global__ void FirstItr_ConnectedComp(CSRGraph graph, DynamicBitset *is_updated
         jj= nps.fg.itvalue[_np_i];
         {
           index_type dst;
-          unsigned long long new_dist;
+          uint32_t new_dist;
           dst = graph.getAbsDestination(jj);
           new_dist = p_comp_current[src];
           if (p_comp_current[dst] > new_dist) {
-            unsigned long long old_dist = atomicMin(&p_comp_current[dst], new_dist);
+            uint32_t old_dist = atomicMin(&p_comp_current[dst], new_dist);
             if (old_dist > new_dist) is_updated->set(dst);
           }
         }
@@ -240,7 +240,7 @@ __global__ void FirstItr_ConnectedComp(CSRGraph graph, DynamicBitset *is_updated
   }
   // FP: "101 -> 102;
 }
-__global__ void ConnectedComp(CSRGraph graph, DynamicBitset *is_updated, unsigned int __nowned, unsigned int __begin, unsigned int __end, unsigned long long * p_comp_current, unsigned long long * p_comp_old, HGAccumulator<int> ret_val)
+__global__ void ConnectedComp(CSRGraph graph, DynamicBitset *is_updated, unsigned int __nowned, unsigned int __begin, unsigned int __end, uint32_t * p_comp_current, uint32_t * p_comp_old, HGAccumulator<int> ret_val)
 {
   unsigned tid = TID_1D;
   unsigned nthreads = TOTAL_THREADS_1D;
@@ -357,11 +357,11 @@ __global__ void ConnectedComp(CSRGraph graph, DynamicBitset *is_updated, unsigne
         jj = ns +_np_j;
         {
           index_type dst;
-          unsigned long long new_dist;
+          uint32_t new_dist;
           dst = graph.getAbsDestination(jj);
           new_dist = p_comp_current[src];
           if (p_comp_current[dst] > new_dist) {
-            unsigned long long old_dist = atomicMin(&p_comp_current[dst], new_dist);
+            uint32_t old_dist = atomicMin(&p_comp_current[dst], new_dist);
             if (old_dist > new_dist) is_updated->set(dst);
           }
         }
@@ -401,11 +401,11 @@ __global__ void ConnectedComp(CSRGraph graph, DynamicBitset *is_updated, unsigne
           jj = _np_w_start +_np_ii;
           {
             index_type dst;
-            unsigned long long new_dist;
+            uint32_t new_dist;
             dst = graph.getAbsDestination(jj);
             new_dist = p_comp_current[src];
             if (p_comp_current[dst] > new_dist) {
-              unsigned long long old_dist = atomicMin(&p_comp_current[dst], new_dist);
+              uint32_t old_dist = atomicMin(&p_comp_current[dst], new_dist);
               if (old_dist > new_dist) is_updated->set(dst);
             }
           }
@@ -441,11 +441,11 @@ __global__ void ConnectedComp(CSRGraph graph, DynamicBitset *is_updated, unsigne
         jj= nps.fg.itvalue[_np_i];
         {
           index_type dst;
-          unsigned long long new_dist;
+          uint32_t new_dist;
           dst = graph.getAbsDestination(jj);
           new_dist = p_comp_current[src];
           if (p_comp_current[dst] > new_dist) {
-            unsigned long long old_dist = atomicMin(&p_comp_current[dst], new_dist);
+            uint32_t old_dist = atomicMin(&p_comp_current[dst], new_dist);
             if (old_dist > new_dist) is_updated->set(dst);
           }
         }
@@ -463,7 +463,7 @@ __global__ void ConnectedComp(CSRGraph graph, DynamicBitset *is_updated, unsigne
   }
   ret_val.thread_exit<_br>(_ts);
 }
-__global__ void ConnectedCompSanityCheck(CSRGraph graph, unsigned int __begin, unsigned int __end, unsigned long long * p_comp_current, HGAccumulator<unsigned int> sum)
+__global__ void ConnectedCompSanityCheck(CSRGraph graph, unsigned int __begin, unsigned int __end, uint32_t * p_comp_current, HGAccumulator<unsigned int> sum)
 {
   unsigned tid = TID_1D;
   unsigned nthreads = TOTAL_THREADS_1D;
